requirements.txt
streamlit
pandas
numpy
scikit-learn
pycaret==3.3.0
matplotlib
seaborn
pillow
